# customer-management-api/target/classes/openapi.yaml
openapi: 3.0.0
info:
  title: Customer Management API
  version: 1.0.0
  description: API to manage customers and calculate their tier

paths:
  /customers:
    post:
      summary: Create a new customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerRequest'
      responses:
        '200':
          description: Customer created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerResponse'

    get:
      summary: Get all customers or search by name/email
      parameters:
        - in: query
          name: name
          schema:
            type: string
        - in: query
          name: email
          schema:
            type: string
      responses:
        '200':
          description: Customers retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CustomerResponse'

  /customers/{id}:
    get:
      summary: Get customer by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Customer found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerResponse'
        '404':
          description: Customer not found

    put:
      summary: Update customer by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerRequest'
      responses:
        '200':
          description: Customer updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerResponse'
        '404':
          description: Customer not found

    delete:
      summary: Delete customer by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Customer deleted successfully
          content:
            text/plain:
              schema:
                type: string
        '404':
          description: Customer not found

components:
  schemas:
    CustomerRequest:
      type: object
      required: [name, email]
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        annualSpend:
          type: number
        lastPurchaseDate:
          type: string
          format: date-time

    CustomerResponse:
      allOf:
        - $ref: '#/components/schemas/CustomerRequest'
      properties:
        id:
          type: string
          format: uuid
        tier:
          type: string